+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 * Add support for GCC.
 * Change Print Extender's `Print` to `ToString` which is a more widely used name.
+* When compiling with Clang show field names with the `AbslFormatImpl` extender.